+++ packages/livedoc-mocha/_src/app/FilterOptions.ts
@@ -1,5 +1,5 @@
 export class FilterOptions {
-    public include: string[] = [];
-    public exclude: string[] = [];
-    public showFilterConflicts: boolean = false;
+    public include?: string[] = [];
+    public exclude?: string[] = [];
+    public showFilterConflicts?: boolean = false;
 }