+++ packages/livedoc-mocha/_src/app/LiveDocOptions.ts
@@ -4,11 +4,11 @@ import { ReporterOptions } from "./reporter";
 import { IPostReporter } from "./reporter/IPostReporter";
 
 export class LiveDocOptions {
-    public constructor () {
+    public constructor() {
     }
-    public rules: LiveDocRules = new LiveDocRules();
-    public filters: FilterOptions = new FilterOptions();
-    public reporterOptions: ReporterOptions = new ReporterOptions();
-    public postReporters: IPostReporter[] = [];
-    isolatedMode: boolean = false;
+    public rules?: LiveDocRules = new LiveDocRules();
+    public filters?: FilterOptions = new FilterOptions();
+    public reporterOptions?: ReporterOptions = new ReporterOptions();
+    public postReporters?: IPostReporter[] = [];
+    public isolatedMode?: boolean = false;
 }