+++ packages/livedoc-mocha/_src/app/model/BddContext.ts
@@ -1,4 +1,4 @@
-import { MochaSuite } from ".";
+import { MochaSuite } from "./MochaSuite";
 
 
 /**