+++ packages/livedoc-mocha/_src/app/reporter/ReporterOptions.ts
@@ -1,5 +1,5 @@
 import { ColorTheme } from "./ColorTheme";
-import { DefaultColorTheme } from ".";
+import { DefaultColorTheme } from "./DefaultColorTheme";
 
 export class ReporterOptions {
     public colors: ColorTheme = new DefaultColorTheme();